cmd/entrypoint: Move the Queue implementation to its own testqueue package
