+++ cmd/entrypoint/internal/testqueue/queue_test.go
@@ -1,15 +1,15 @@
-package entrypoint_test
+package testqueue_test
 
 import (
 	"testing"
 	"time"
 
-	"github.com/datawire/ambassador/v2/cmd/entrypoint"
+	"github.com/datawire/ambassador/v2/cmd/entrypoint/internal/testqueue"
 	"github.com/stretchr/testify/require"
 )
 
 func TestFakeQueueGet(t *testing.T) {
-	q := entrypoint.NewQueue(t, 10*time.Second)
+	q := testqueue.NewQueue(t, 10*time.Second)
 
 	go func() {
 		for count := 0; count < 10; count++ {
@@ -27,7 +27,7 @@ func TestFakeQueueGet(t *testing.T) {
 }
 
 func TestFakeQueueSkip(t *testing.T) {
-	q := entrypoint.NewQueue(t, 10*time.Second)
+	q := testqueue.NewQueue(t, 10*time.Second)
 
 	go func() {
 		for count := 0; count < 10; count++ {
